+++ pages/RamaeraRouter.tsx
@@ -1,15 +1,14 @@
-import { GET_AGENCY_CODE, GetUser } from '@/apollo/queries/auth';
+import { GET_AGENCY_CODE, GET_ALL_USERS, GetUser } from '@/apollo/queries/auth';
 import { useAppDispatch } from '@/hooks';
+import { setAllTheUsers, gotData } from '@/state/slice/allUsersSlice';
 import { setAgencyCode, setOrUpdateUser } from '@/state/slice/userSlice';
 import { useQuery } from '@apollo/client';
 import { useRouter } from 'next/router';
 import { useEffect, useState } from 'react';
-import { useSelector } from 'react-redux';
 
 const RamaeraRouter = ({ children }) => {
   const router = useRouter();
   const userResp = useQuery(GetUser);
-
   const agencyCode = useQuery(GET_AGENCY_CODE, {
     variables: { userID: userResp?.data?.me.id }
   });
