+++ pages/_app.tsx
@@ -15,11 +15,11 @@ import Router from 'next/router';
 import nProgress from 'nprogress';
 import 'nprogress/nprogress.css';
 import { Provider } from 'react-redux';
+import { PersistGate } from 'redux-persist/integration/react';
 import { SidebarProvider } from 'src/contexts/SidebarContext';
 import createEmotionCache from 'src/createEmotionCache';
 import ThemeProvider from 'src/theme/ThemeProvider';
 import { persistor, store } from '../src/state/store';
-import { PersistGate } from 'redux-persist/integration/react';
 import RamaeraRouter from './RamaeraRouter';
 import './style.css';
 library.add(faEye, faEyeSlash);
