+++ pages/dashboard/project/[index].tsx
@@ -5,12 +5,65 @@ import Head from 'next/head';
 import { useRouter } from 'next/router';
 import ProtectedSSRoute from 'pages/libs/ProtectedRoute';
 import PaymentTab from './Tabs/Payment';
+import { useDispatch, useSelector } from 'react-redux';
+import {
+  gotData,
+  setAllTheUsers,
+  setTotalAgraAmount,
+  setTotalHajipurAmount
+} from '@/state/slice/allUsersSlice';
+import { useEffect } from 'react';
+import { GET_ALL_USERS } from '@/apollo/queries/auth';
+import { useQuery } from '@apollo/client';
 
 function index() {
   const router = useRouter();
+  const dispatch = useDispatch();
   let title =
     router.query.index.charAt(0).toUpperCase() + router.query.index.slice(1);
 
+  let usersList = useSelector((state: any) => state.allUsers.allTheUsers);
+
+  const getAllUser = useQuery(GET_ALL_USERS);
+  if (getAllUser.data && !usersList[0]) {
+    dispatch(setAllTheUsers(getAllUser.data.getAllUser));
+    dispatch(gotData(true));
+  }
+
+  const setProjectAmount = async () => {
+    let hajipurAmount = 0;
+    let agraAmount = 0;
+    if (usersList[0]) {
+      await usersList.map((user) => {
+        user?.documents.map((doc) => {
+          doc.title.includes('hajipur') && (hajipurAmount += doc.amount);
+          doc.title.includes('agra') && (agraAmount += doc.amount);
+        });
+      });
+      await dispatch(setTotalHajipurAmount(hajipurAmount));
+      await dispatch(setTotalAgraAmount(agraAmount));
+    }
+  };
+
+  useEffect(() => {
+    setProjectAmount();
+  }, [usersList]);
+
+  if (!usersList[0]) {
+    return (
+      <div
+        style={{
+          display: 'flex',
+          justifyContent: 'center',
+          height: '90vh',
+          alignItems: 'center'
+        }}
+      >
+        <h2>Loading...</h2>
+        <img src="" />
+      </div>
+    );
+  }
   return (
     <ProtectedSSRoute>
       <Head>
