+++ pages/list/index.tsx
@@ -4,8 +4,21 @@ import Head from 'next/head';
 import { Container } from '@mui/material';
 import ProtectedSSRoute from 'pages/libs/ProtectedRoute';
 import UserTable from './components/UserTable';
+import { useDispatch, useSelector } from 'react-redux';
+import { GET_ALL_USERS } from '@/apollo/queries/auth';
+import { setAllTheUsers, gotData } from '@/state/slice/allUsersSlice';
+import { useQuery } from '@apollo/client';
 
 function index() {
+  let usersList = useSelector((state: any) => state.allUsers.allTheUsers);
+
+  const dispatch = useDispatch();
+  const getAllUser = useQuery(GET_ALL_USERS);
+  if (getAllUser.data && !usersList[0]) {
+    dispatch(setAllTheUsers(getAllUser.data.getAllUser));
+    dispatch(gotData(true));
+  }
+
   return (
     <ProtectedSSRoute>
       <Head>
